requirements.txt
aio-pika==9.5.4
aiormq==6.8.1
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.8
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
cachetools==5.5.0
carehare==1.0.2
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels-rabbitmq==4.0.1
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
dill==0.3.9
Django==5.1.4
django-cors-headers==4.6.0
django-stubs-ext==5.1.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
exceptiongroup==1.2.2
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
kombu==5.4.2
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pamqp==3.3.0
pathspec==0.12.1
pika==1.3.2
pillow==11.1.0
pip==24.3.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.2
Twisted==24.11.0
txaio==23.1.1
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
zope.interface==7.2
